+++ tests/changes.rs
@@ -2,7 +2,7 @@ use std::io::Cursor;
 
 use bevy::{ecs::entity::MapEntities, prelude::*, utils::Duration};
 use bevy_replicon::{
-    client::{confirmed::Confirmed, server_entity_map::ServerEntityMap, ServerInitTick},
+    client::{confirm_history::ConfirmHistory, server_entity_map::ServerEntityMap, ServerInitTick},
     core::{
         command_markers::MarkerConfig,
         ctx::WriteCtx,
@@ -449,11 +449,11 @@ fn marker_with_history_old_update() {
     // so that the next update for this entity is discarded.
     let mut tick = **server_app.world.resource::<ServerTick>();
     tick += u64::BITS + 1;
-    let mut confirmed = client_app
+    let mut confirm_history = client_app
         .world
-        .get_mut::<Confirmed>(client_entity)
+        .get_mut::<ConfirmHistory>(client_entity)
         .unwrap();
-    confirmed.confirm(tick);
+    confirm_history.confirm(tick);
 
     let mut component = server_app
         .world